refactor: nest project components inside folder
