+++ components/projects/expenses-tracker/ExpensesChart.tsx
@@ -1,9 +1,9 @@
 //? Default styled header
-import { StyledSubHeader } from "../../components/UI/StyledSubHeader.tsx";
+import { StyledSubHeader } from "../../../components/UI/StyledSubHeader.tsx";
 //? Import individual MonthBar components to render all 12 months
 import { ExpensesChartMonthBar } from "./ExpensesChartMonthBar.tsx";
 //? Import the Expense type for casting
-import type { Expense } from "../../types/expenses-tracker/Expense.ts";
+import type { Expense } from "../../../types/expenses-tracker/Expense.ts";
 
 //? Define properties for this component
 interface ExpenseChartProperties {
