+++ islands/ExpensesTracker.tsx
@@ -1,13 +1,13 @@
 //? Import from Preact to be able to change state
 import { useState } from "preact/hooks";
 //? Import IndividualExpenses component to display expenses
-import { IndividualExpense } from "../components/expenses-tracker/IndividualExpense.tsx";
+import { IndividualExpense } from "../components/projects/expenses-tracker/IndividualExpense.tsx";
 //? Import expense creation form
 import AddNewExpense from "./AddNewExpense.tsx";
 //? Import the Expense type for casting
 import type { Expense } from "../types/expenses-tracker/Expense.ts";
 import ExpensesYearSelect from "./ExpensesYearSelect.tsx";
-import { ExpenseChart } from "../components/expenses-tracker/ExpensesChart.tsx";
+import { ExpenseChart } from "../components/projects/expenses-tracker/ExpensesChart.tsx";
 import { postNewExpense } from "../services/expenses-tracker/postNewExpense.ts";
 
 //? Filters expenses by year
