+++ islands/FoodOrder.tsx
@@ -3,13 +3,13 @@ import { useEffect, useRef, useState } from "preact/hooks";
 //? Import Food type to typecast the data received
 import type { Food } from "../types/food-order/Food.ts";
 //? Renders invidual food items on the page
-import { FoodItem } from "../components/food-order/FoodItem.tsx";
+import { FoodItem } from "../components/projects/food-order/FoodItem.tsx";
 //? Creates a modal with a shaded/blurred backdrop over the content behind it
 import { ModalWithBackdrop } from "../components/UI/ModalWithBackdrop.tsx";
 //? Cart Button that opens the current Cart as a Modal
-import CartButton from "../components/food-order/CartButton.tsx";
+import CartButton from "../components/projects/food-order/CartButton.tsx";
 //? Content of the current Cart in a modal for the user to view Cart Items
-import CartModal from "../components/food-order/CartModal.tsx";
+import CartModal from "../components/projects/food-order/CartModal.tsx";
 //? Import accent button to create the Order now! button at the bottom
 import { AccentButton } from "../components/UI/AccentButton.tsx";
 import { ModalExtendedImage } from "../components/UI/ModalExtendedImage.tsx";
